tests/api: Split user tests
... and use "john" instead of "test_user"
